Improve prepare_package script

+++ test/integration/commands/prepare-package.sh
@@ -1,14 +1,22 @@
 #!/usr/bin/env bash
 
-NARVAL_FOO_NODE_MODULES="test/integration/packages/api/_node_modules"
+BASE_PATH="test/integration/packages"
+NARVAL_FOO_NODE_MODULES="$BASE_PATH/api/_node_modules"
 NARVAL_LIB_DEST="${NARVAL_FOO_NODE_MODULES}/narval"
-NARVAL_FILE="test/integration/packages/${package_to_launch}/.narval.yml"
-COVERAGE_FOLDER="test/integration/packages/${package_to_launch}/.coverage"
+NARVAL_FILE="$BASE_PATH/${package_to_launch}/.narval.yml"
+COVERAGE_FOLDER="$BASE_PATH/${package_to_launch}/.coverage"
+SHARED_FOLDER="$BASE_PATH/${package_to_launch}/.shared"
 LOG_SEP=">>>>>>>>>>>>>>>"
 
 echo "$LOG_SEP CLEANING \"${package_to_launch}\" PACKAGE"
 # sudo needed if docker write files with root user, as in Travis builds
-rm -rf $COVERAGE_FOLDER || sudo rm -rf $COVERAGE_FOLDER
+if [ -d $COVERAGE_FOLDER ]; then
+  rm -rf $COVERAGE_FOLDER || sudo rm -rf $COVERAGE_FOLDER
+fi
+
+if [ -d $SHARED_FOLDER ]; then
+  rm -rf $SHARED_FOLDER
+fi
 
 if [ -f $NARVAL_FILE ]; then
   rm $NARVAL_FILE
